--- a/axelrod/__init__.py
+++ b/axelrod/__init__.py
@@ -1,8 +1,11 @@
 from plot import *
 from game import *
 from player import *
+from mock_player import *
 from round_robin import *
 from tournament import *
 from tournament_manager import *
 from strategies import *
 from ecosystem import *
+
+