--- a/axelrod/strategies/_strategies.py
+++ b/axelrod/strategies/_strategies.py
@@ -44,6 +44,7 @@ ordinary_strategies = [
     AverageCopier,
     BackStabber,
     Bully,
+    Calculator,
     CautiousQLearner,
     Champion,
     CooperatorHunter,