--- a/axelrod/strategies/_strategies.py
+++ b/axelrod/strategies/_strategies.py
@@ -55,12 +55,12 @@ ordinary_strategies = [
     AlternatorHunter,
     Appeaser,
     AntiTitForTat,
-    ArrogantQLearner,
+    #ArrogantQLearner, # needs fixing
     AverageCopier,
     BackStabber,
     Bully,
     #Calculator,
-    CautiousQLearner,
+    #CautiousQLearner, # needs fixing
     #Champion,
     CooperatorHunter,
     #Davis,
@@ -86,7 +86,7 @@ ordinary_strategies = [
     HardProber,
     HardTitForTat,
     HardTitFor2Tats,
-    HesitantQLearner,
+    #HesitantQLearner, # needs fixing
     Inverse,
     InversePunisher,
     Joss,
@@ -110,7 +110,7 @@ ordinary_strategies = [
     Retaliate,
     Retaliate2,
     Retaliate3,
-    RiskyQLearner,
+    #RiskyQLearner, # Needs fixing
     #Shubik,
     SoftJoss,
     SneakyTitForTat,