--- a/axelrod/strategies/appeaser.py
+++ b/axelrod/strategies/appeaser.py
@@ -12,12 +12,16 @@ class Appeaser(Player):
     memory_depth = float('inf')  # Depends on internal memory.
 
     def strategy(self, opponent):
-        if len(self.history) == 0:
-            self.str = 'C'
+        if not len(self.history):
+            self.play = 'C'
         else:
             if opponent.history[-1] == 'D':
-                if self.str == 'C':
-                    self.str = 'D'
+                if self.play == 'C':
+                    self.play = 'D'
                 else:
-                    self.str = 'C'
-        return self.str
+                    self.play = 'C'
+        return self.play
+
+    def reset(self):
+        Player.reset(self)
+        self.play = 'C'