--- a/axelrod/strategies/averagecopier.py
+++ b/axelrod/strategies/averagecopier.py
@@ -13,12 +13,11 @@ class AverageCopier(Player):
         """Randomly picks a strategy (not affected by history)."""
         if len(opponent.history) == 0:
             return random.choice(['C', 'D'])
-        p = sum([s == 'C' for s in opponent.history]) / len(opponent.history)
+        p = opponent.cooperations / len(opponent.history)
         rnd_num = random.random()
         if rnd_num < p:
             return 'C'
         return 'D'
-
 
 class NiceAverageCopier(Player):
     """Same as Average Copier, but always starts by cooperating."""
@@ -31,7 +30,7 @@ class NiceAverageCopier(Player):
         """Randomly picks a strategy (not affected by history)."""
         if len(opponent.history) == 0:
             return 'C'
-        p = sum([s == 'C' for s in opponent.history]) / len(opponent.history)
+        p = opponent.cooperations / len(opponent.history)
         rnd_num = random.random()
         if rnd_num < p:
             return 'C'