--- a/axelrod/strategies/defector.py
+++ b/axelrod/strategies/defector.py
@@ -24,6 +24,6 @@ class TrickyDefector(Player):
         Defect if opponent has cooperated at least once in the past and has defected
         for the last 3 turns in a row.
         """
-        if 'C' in opponent.history and opponent.history[-3:] == ['D']*3:
+        if 'C' in opponent.history and opponent.history[-3:] == ['D'] * 3:
             return 'C'
         return 'D'