--- a/axelrod/strategies/forgiver.py
+++ b/axelrod/strategies/forgiver.py
@@ -16,7 +16,7 @@ class Forgiver(Player):
         Begins by playing C, then plays D if the opponent has defected more than 10 percent of the time
         """
         try:
-            if opponent.history.count('D') > len(opponent.history) / 10.:
+            if opponent.defections > len(opponent.history) / 10.:
                 return 'D'
             return 'C'
         except IndexError:
@@ -41,7 +41,7 @@ class ForgivingTitForTat(Player):
         and their most recent decision was defect.
         """
         try:
-            if opponent.history.count('D') > len(opponent.history) / 10.:
+            if opponent.defections > len(opponent.history) / 10.:
                 return opponent.history[-1]
             return 'C'
         except IndexError: