--- a/axelrod/strategies/grumpy.py
+++ b/axelrod/strategies/grumpy.py
@@ -14,7 +14,7 @@ class Grumpy(Player):
         """Player starts of nice be default with set thresholds"""
         super(Grumpy, self).__init__()
         self.history = []
-        self.score = 0
+        #self.score = 0
         self.state = starting_state
         self.starting_state = starting_state
         self.grumpy_threshold = grumpy_threshold
@@ -46,6 +46,5 @@ class Grumpy(Player):
 
     def reset(self):
         """Resets score, history and state for the next round of the tournement."""
-
-        self.history = []
+        Player.reset(self)
         self.state = self.starting_state