--- a/axelrod/strategies/mathematicalconstants.py
+++ b/axelrod/strategies/mathematicalconstants.py
@@ -10,16 +10,18 @@ class CotoDeRatio(Player):
 
     def strategy(self, opponent):
 
-        """initially co-operates"""
+        # Initially cooperate
         if len(opponent.history) == 0:
             return 'C'
 
-        """to avoid initial division by zero"""
-        if sum([s == 'D' for s in opponent.history]) == 0:
+        # Avoid initial division by zero
+        if not opponent.defections:
             return 'D'
 
-        """otherwise compare ratio to golden mean"""
-        if (sum([p == 'C' for p in opponent.history]) + sum([q == 'C' for q in self.history]))/(sum([x == 'D' for x in opponent.history]) + sum([y == 'D' for y in self.history])) > self.ratio:
+        # Otherwise compare ratio to golden mean
+        cooperations = opponent.cooperations + self.cooperations
+        defections = float(opponent.defections + self.defections)
+        if cooperations / defections > self.ratio:
             return 'D'
         return 'C'