… axelrod/strategies/cooperator.py
… axelrod/strategies/oncebitten.py
… axelrod/strategies/punisher.py
… axelrod/strategies/qlearner.py
… axelrod/strategies/retaliate.py
… axelrod/strategies/titfortat.py
… axelrod/tests/unit/test_alternator.py
… axelrod/tests/unit/test_appeaser.py
… axelrod/tests/unit/test_axelrod_tournaments.py
… axelrod/tests/unit/test_backstabber.py
… axelrod/tests/unit/test_calculator.py
… axelrod/tests/unit/test_cooperator.py
… axelrod/tests/unit/test_defector.py
… axelrod/tests/unit/test_forgiver.py
… axelrod/tests/unit/test_game.py
… axelrod/tests/unit/test_gobymajority.py
… axelrod/tests/unit/test_grudger.py
… axelrod/tests/unit/test_grumpy.py
… axelrod/tests/unit/test_headsup.py
… axelrod/tests/unit/test_hunter.py
… axelrod/tests/unit/test_mathematicalconstants.py
… axelrod/tests/unit/test_meta.py
… axelrod/tests/unit/test_mindcontrol.py
… axelrod/tests/unit/test_mindreader.py
… axelrod/tests/unit/test_mock_player.py
… axelrod/tests/unit/test_oncebitten.py
… axelrod/tests/unit/test_player.py
… axelrod/tests/unit/test_punisher.py
… axelrod/tests/unit/test_retaliate.py