Many strategies were scanning either the opponents or their own histories each time `strategy()` was called, resulting in about `O(tournament_length^2)` unnecessary operations in every matchup versus simply tracking the total number of cooperations and defections. So I added automatic counting of cooperations and defections to the `Player` class and modified many strategies accordingly.

This reduces the time to run all the tests to 50 seconds on my machine (despite the fact that I added many new tests as well), down from 135 seconds. Along the way I made a number of improvements to the existing strategies and tests.

Also, since the Qlearners are bugged, I commented them out of `_strategies.py`. They use the wrong game matrix and all the tests fail when the correct game matrix is inserted.

Finally, the suggestion in #255 only slightly improves matters (50 seconds down to 40 seconds for all tests), so I closed it.