@meatballs 2015-08-06
An excellent idea! I'm more than happy for this to go in.