@drvinceknight 2015-08-06
> Also, since the Qlearners are bugged, I commented them out of _strategies.py. They use the wrong game matrix and all the tests fail when the correct game matrix is inserted.

:+1: Great call. Tagging @geraintpalmer so that he is aware: #157